Wrap into a package
