+++ src/main/java/carbon/task/Deadline.java
@@ -1,35 +1,39 @@
-package carbon;
+package carbon.task;
 
+import carbon.error.CarbonException;
+import carbon.error.InvalidFlagException;
+import carbon.error.InvalidParamException;
+import carbon.error.InvalidTimeException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
 import java.time.temporal.Temporal;
 
-class Event extends Task {
-    public static final String FLAG = " /at";
-    private static final int TYPEKEY = Task.getTypeKey(Task.Type.EVENT);
+public class Deadline extends Task {
+    public static final String FLAG = " /by";
+    private static final int TYPEKEY = Task.getTypeKey(Task.Type.DEADLINE);
 
-    private static String extractName(String input) {
-        int flagIndex = input.indexOf(Event.FLAG);
+    private static String extractName(String input) throws CarbonException {
+        int flagIndex = input.indexOf(Deadline.FLAG);
         if (flagIndex == -1) {
-            CarbonException invalidFlag = new InvalidFlagException(input, "event");
+            CarbonException invalidFlag = new InvalidFlagException(input, "deadline");
             throw invalidFlag;
         } else {
-            String name = input.substring("event ".length(), flagIndex);
+            String name = input.substring("deadline ".length(), flagIndex);
             return name;
         }
     }
 
     private static Temporal extractTime(String input) throws CarbonException {
         int len = input.length();
-        int flagIndex = input.indexOf(Event.FLAG);
-        if (len <= flagIndex + Event.FLAG.length() + 1) {
+        int flagIndex = input.indexOf(Deadline.FLAG);
+        if (len <= flagIndex + Deadline.FLAG.length() + 1) {
             CarbonException invalidParam = new InvalidParamException(input);
             throw invalidParam;
         } else {
-            String timeString = input.substring(flagIndex + Event.FLAG.length() + 1);
+            String timeString = input.substring(flagIndex + Deadline.FLAG.length() + 1);
             try {
-                Temporal dateTime = Event.formatTime(timeString);
+                Temporal dateTime = Deadline.formatTime(timeString);
                 return dateTime;
             } catch (DateTimeParseException error) {
                 CarbonException invalidTime = new InvalidTimeException(timeString);
@@ -54,30 +58,29 @@ class Event extends Task {
         }
     }
 
-
     public static Task createTask(String input) {
-        String name = Event.extractName(input);
-        Temporal dateTime = Event.extractTime(input);
-        Task event = new Event(name, false, dateTime);
-        return event;
+        String name = Deadline.extractName(input);
+        Temporal dateTime = Deadline.extractTime(input);
+        Task deadline = new Deadline(name, false, dateTime);
+        return deadline;
     }
 
     public static Task loadTask(String name, Boolean isDone, String time) {
-        Temporal dateTime = Event.formatTime(time);
-        Task event = new Event(name, isDone, dateTime);
-        return event;
+        Temporal dateTime = Deadline.formatTime(time);
+        Task deadline = new Deadline(name, isDone, dateTime);
+        return deadline;
     }
 
     private Temporal dateTime;
-
-    private Event(String name, Boolean isDone, Temporal dateTime) {
+    
+    private Deadline(String name, Boolean isDone, Temporal dateTime) {
         super(name, isDone);
         this.dateTime = dateTime;
     }
 
     @Override
     public String encode() {
-        int typeKey = Event.TYPEKEY;
+        int typeKey = Deadline.TYPEKEY;
         int isDone = this.isDone ? 1 : 0;
         String result = String.format(
                 "%d|%d|%s|%s\n", typeKey, isDone, this.name, 
@@ -95,16 +98,16 @@ class Event extends Task {
         } else {
             LocalDateTime time = (LocalDateTime) this.dateTime;
             timeFormatted = formatType == Task.FormatType.READ
-                    ? time.format(Task.dateFormat)
-                    : time.format(Task.dateFormatPrint);
+                    ? time.format(Task.dateTimeFormat)
+                    : time.format(Task.dateTimeFormatPrint);
         }
         return timeFormatted;
     }
 
     @Override
     public String toString() {
-        String type = "\u001B[34m(EVNT)\u001B[0m";
+        String type = "\u001B[32m(DEAD)\u001B[0m";
         String timeFormatted = this.formatTime(Task.FormatType.PRINT);
-        return String.format("%s %s @ %s", type, super.toString(), timeFormatted);
+        return String.format("%s %s < %s", type, super.toString(), timeFormatted);
     }
 }
